+++ src/js/lib/pages/github/issue.js
@@ -8,54 +8,21 @@ import K2pickerarea from '../../../module/K2pickerarea/K2pickerarea';
 import K2pickerType from '../../../module/K2pickertype/K2pickertype';
 import ToggleReview from '../../../module/ToggleReview/ToggleReview';
 import K2comments from '../../../module/K2comments/K2comments';
+import K2previousissues from '../../../module/K2previousissues/K2previousissues';
 import ONYXKEYS from '../../../ONYXKEYS';
 import * as API from '../../api';
 
 let clearErrorTimeoutID;
 function catchError(e) {
-    $('.gh-header-actions .k2-element').remove();
-    $('.gh-header-actions').append('<span class="alert k2-element">OOPS!</span>');
+    $('div[data-component="PH_Actions"] .k2-element').remove(); // K2 elements in action buttons
+    $('div[data-component="PH_Actions"]') // Action buttons next to issue title
+        .append('<span class="alert k2-element">OOPS!</span>');
     console.error(e);
     clearTimeout(clearErrorTimeoutID);
     clearErrorTimeoutID = setTimeout(() => {
-        $('.gh-header-actions .k2-element').remove();
+        $('div[data-component="PH_Actions"] .k2-element').remove();
     }, 30000);
 }
-
-/**
- * Gets the contents of the reviewer checklist from GitHub and then posts it as a comment to the current PR
- * @param {Event} e
- */
-const copyReviewerChecklist = (e) => {
-    e.preventDefault();
-    const pathToChecklist = 'https://raw.githubusercontent.com/Expensify/App/main/contributingGuides/BUGZERO_CHECKLIST.md';
-    $.get(pathToChecklist)
-        .done((fileContents) => {
-            if (!fileContents) {
-                console.error(`could not load contents of ${pathToChecklist} for some reason`);
-                return;
-            }
-
-            API.addComment(fileContents);
-        });
-};
-
-const renderCopyChecklistButton = () => {
-    // Look through all the comments on the page to find one that has the template for the copy/paste checklist button
-    // eslint-disable-next-line rulesdir/prefer-underscore-method
-    $('.js-comment-body').each((i, el) => {
-        const commentHtml = $(el).html();
-
-        // When the button template is found, replace it with an HTML button and then put that back into the DOM so someone can click on it
-        if (commentHtml && commentHtml.indexOf('you can simply click: [this button]') > -1) {
-            const newHtml = commentHtml.replace('[this button]', '<button type="button" class="btn btn-sm k2-copy-checklist">HERE</button>');
-            $(el).html(newHtml);
-
-            // Now that the button is on the page, add a click handler to it (always remove all handlers first so that we know there will always be one handler attached)
-            $('.k2-copy-checklist').off().on('click', copyReviewerChecklist);
-        }
-    });
-};
 
 /**
  * Sets the owner of an issue when it doesn't have an owner yet
@@ -104,27 +71,33 @@ function replaceOwner(oldOwner, newOwner) {
 
 /**
  * This method is all about adding the "issue owner" functionality which melvin will use to see who should be providing ksv2 updates to an issue.
+ * @param {String | null} [issueOwner] GitHub username of the issue owner. Null means no owner, undefined means get the owner from issue body
  */
-const refreshAssignees = () => {
+const renderAssignees = (issueOwner) => {
     // Always start by erasing whatever was drawn before (so it always starts from a clean slate)
-    $('.js-issue-assignees .k2-element').remove();
+    $('div[data-testid="sidebar-section"] .k2-element').remove();
 
-    // Check if there is an owner for the issue
-    const ghDescription = $('.comment-body').text();
-    const regexResult = ghDescription.match(/Current Issue Owner:\s@(?<owner>\S+)/i);
-    const currentOwner = regexResult && regexResult.groups && regexResult.groups.owner;
+    let currentOwner = issueOwner;
+
+    // if issue owner is not provided, then try to get it from the issue body
+    if (currentOwner === undefined) {
+        const ghDescription = $('.markdown-body').first().text();
+        const regexResult = ghDescription.match(/Current Issue Owner:\s@(?<owner>\S+)/i);
+        currentOwner = regexResult && regexResult.groups && regexResult.groups.owner;
+    }
 
     // Add buttons to each assignee
-    $('.js-issue-assignees > p > span').each((i, el) => {
-        const assignee = $(el).find('.assignee span').text();
+    $('div[data-testid="issue-assignees"]').each((i, el) => {
+        const assignee = $(el).text();
+        $(el).closest('li').css('display', 'flex');
         if (assignee === currentOwner) {
-            $(el).append(`
+            $(el).closest('li').append(`
                 <button type="button" class="Button flex-md-order-2 m-0 owner k2-element k2-button k2-button-remove-owner" data-owner="${currentOwner}">
                     ★
                 </button>
             `);
         } else {
-            $(el).append(`
+            $(el).closest('li').append(`
                 <button type="button" class="Button flex-md-order-2 m-0 k2-element k2-button k2-button-make-owner" data-owner="${assignee}">
                     ☆
                 </button>
@@ -133,15 +106,15 @@ const refreshAssignees = () => {
     });
 
     // Remove the owner with this button is clicked
-    $('.k2-button-remove-owner').off('click').on('click', (e) => {
+    $('.k2-button-remove-owner').off('click').on('click', async (e) => {
         e.preventDefault();
         const owner = $(e.target).data('owner');
         removeOwner(owner);
-        return false;
+        renderAssignees(null);
     });
 
     // Make a new owner when this button is clicked
-    $('.k2-button-make-owner').off('click').on('click', (e) => {
+    $('.k2-button-make-owner').off('click').on('click', async (e) => {
         e.preventDefault();
         const newOwner = $(e.target).data('owner');
         if (currentOwner) {
@@ -149,14 +122,15 @@ const refreshAssignees = () => {
         } else {
             setOwner(newOwner);
         }
-        return false;
+        renderAssignees(newOwner);
     });
 };
 
 const refreshPicker = function () {
     // Add our wrappers to the DOM which all the React components will be rendered into
-    if (!$('.k2picker-wrapper').length) {
-        $('.js-issue-labels').after(sidebarWrapperHTML);
+    if ($('.k2picker-wrapper')) {
+        $('div[data-testid="issue-viewer-metadata-pane"] > :nth-child(4)') // Labels section in right side panel
+            .after(sidebarWrapperHTML);
     }
 
     new K2picker().draw();
@@ -164,6 +138,7 @@ const refreshPicker = function () {
     new K2pickerarea().draw();
     new ToggleReview().draw();
     new K2comments().draw();
+    new K2previousissues().draw();
 };
 
 /**
@@ -172,7 +147,7 @@ const refreshPicker = function () {
  * @returns {Object}
  */
 export default function () {
-    let allreadySetup = false;
+    let alreadySetup = false;
     ReactNativeOnyx.init({
         keys: ONYXKEYS,
     });
@@ -183,27 +158,29 @@ export default function () {
 
     IssuePage.setup = function () {
         // Prevent this function from running twice (it sometimes does that because of how chrome triggers the extension)
-        if (allreadySetup) {
+        if (alreadySetup) {
             return;
         }
-        allreadySetup = true;
+        alreadySetup = true;
 
         // Draw them once when the page is loaded
         setTimeout(refreshPicker, 500);
-        setTimeout(refreshAssignees, 500);
+        setTimeout(renderAssignees, 500);
 
         // Every second, check to see if the pickers are still there, and if not, redraw them
         setInterval(() => {
-            if (!$('.k2picker-wrapper').length) {
+            if (!$('.k2picker-wrapper')) {
                 refreshPicker();
             }
-            if (!$('.js-issue-assignees .k2-element').length) {
-                refreshAssignees();
+
+            if (!$('div[data-testid="issue-viewer-metadata-pane"] > :nth-child(3) .k2-element') // Assignee section in right side panel
+                .length) {
+                renderAssignees();
             }
         }, 1000);
 
         // Waiting 2 seconds to call this gives the page enough time to load so that there is a better chance that all the comments will be rendered
-        setInterval(renderCopyChecklistButton, 2000);
+        setInterval(() => IssuePage.renderCopyChecklistButtons('bugzero'), 2000);
     };
 
     return IssuePage;